ZadanieEshopApi/requirements.txt
attrs==25.1.0
black==25.1.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
flake8==7.1.2
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mccabe==0.7.0
mistune==3.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pymongo==4.11.2
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
